requirements.txt
torch==0.4.1
torchvision==0.2.1
tensorboardX
pillow
numpy
#opencv-contrib-python